Created demo of how changing screens should work
Created a demo that creates a p5 object, p5A.

When the user clicks the mouse, p5A receives that mouse click event and
deletes itself, deletes its canvas, and instantiates a p5 object called
p5B.

We can use this same approach to create a separate menu screen and a
separate game screen. The functionality for each of these can be defined
in separate Javascript files.